推送：修改了商店部分功能

+++ src/main/java/mapper/ShopProfileMapper.java
@@ -8,5 +8,5 @@ public interface ShopProfileMapper {
     ShopProfile findById(@Param("userId") String userId);
     int insert(ShopProfile shopProfile);
     int update(ShopProfile shopProfile);
-    int updateBalance(@Param("userId") String userId, @Param("newBalance") BigDecimal newBalance);
+    int updateBalance(@Param("userId") String userId, @Param("newBalance") Double newBalance);
 }
